requirements.txt
fastapi
uvicorn
pyyaml
requests
beautifulsoup4
